+++ lib/src/test/java/org/eclipse/tsp/java/client/api/bookmark/BookmarkApiTest.java
@@ -1,0 +1,13 @@
+package org.eclipse.tsp.java.client.api.bookmark;
+
+import com.github.tomakehurst.wiremock.junit5.WireMockTest;
+
+@WireMockTest(httpPort = 8080)
+public class BookmarkApiTest {
+
+    private static final String FIXTURE_PATH = "fixtures/tspclient";
+    private static final String TSP_EXTENSION_URL = "/tsp/api";
+
+    private BookmarkApi bookmarkApi = new BookmarkApi("http://localhost:8080");
+
+}
